+++ docs/docs.yml
@@ -407,8 +407,6 @@ navigation:
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyFeeEstimator.mdx
                   - page: alchemyGasAndPaymasterAndDataMiddleware
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyGasAndPaymasterAndDataMiddleware.mdx
-                  - page: alchemyGasManagerHooks
-                    path: wallets/pages/reference/account-kit/infra/functions/alchemyGasManagerHooks.mdx
                   - page: alchemyGasManagerMiddleware
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyGasManagerMiddleware.mdx
                   - page: alchemyUserOperationSimulator
@@ -1161,6 +1159,8 @@ navigation:
             contents:
               - section: Functions
                 contents:
+                  - page: alchemyGasManagerHooks
+                    path: wallets/pages/reference/alchemy/smart-accounts/functions/alchemyGasManagerHooks.mdx
                   - page: defaultLightAccountVersion
                     path: wallets/pages/reference/alchemy/smart-accounts/functions/defaultLightAccountVersion.mdx
                   - page: getDefaultLightAccountFactoryAddress
