+++ docs/pages/reference/alchemy/smart-accounts/functions/alchemyGasManagerHooks.mdx
@@ -2,7 +2,7 @@
 # This file is autogenerated
 title: alchemyGasManagerHooks
 description: Overview of the alchemyGasManagerHooks method
-slug: wallets/reference/account-kit/infra/functions/alchemyGasManagerHooks
+slug: wallets/reference/alchemy/smart-accounts/functions/alchemyGasManagerHooks
 ---
 
 Adapts Alchemy's Gas Manager to viem's paymaster interface using the optimized
@@ -15,14 +15,14 @@ when viem calls both getPaymasterStubData and getPaymasterData.
 ## Import
 
 ```ts
-import { alchemyGasManagerHooks } from "@account-kit/infra";
+import { alchemyGasManagerHooks } from "@alchemy/smart-accounts";
 ```
 
 ## Usage
 
 ```ts
 import { createBundlerClient } from "viem/account-abstraction";
-import { alchemyGasManagerHooks } from "@account-kit/infra";
+import { alchemyGasManagerHooks } from "@alchemy/smart-accounts";
 
 const bundler = createBundlerClient({
   transport: http("https://eth-sepolia.g.alchemy.com/v2/your-api-key"),
